Hangfire 1.7 support (#17)
* Updated packages and compilation targets

* Replaced obsolete methods

* Replaced obsolete methods

* Consolidated packages

* Removed outdated properties

* Formatting and beautifying

* Renamed metric

* Rewritten test

* Refactoring

* Various improvements

* Fixed tests

* Replaced with nameof

* Refactoring

- Maintenance processes use Wait() instead of WaitHandle.WaitOne()
- Storage code split into meaning parts
- Implemented Execute for ConnectionHolder

* Refactoring

* Used new helper methods

* Fixed tests

* Updated dependencies

* Queue ordering experiment

* Migrated ids to long

* Next version will be 1.7

* Project file cleanup + test target added

* Adjusted metrics name

To be more stylish :)

* Counters aggregation is made much often

* Perf improvement to monitoring api

* Fixed migration

* Added documentation tags

* Added missing method

* Rewritten monitoring query

* Updated migration

* Refactoring

* Fixed hash count

* Refactored expiration manager tests

* Tests, documentation

* Job creation done in transaction

* All SQL code are converted to lowercase

* Respect postgres server time

* Added test

* Just experimeting

* Refactored tests

* Tests refactoring

* Added tests for monitoring api

* Added tests

* More tests for monitoring api

* Temporarily switched to single runtime

* Removed unused field

* Refactoring

* Refactoring

* Refactoring

* Refactoring

* Optimized monitoring query

now involves only one index scan

* Fetched job made sealed

* Refactoring

* fixed: respect jobqueue order

* Added integration test Hangfire.Console

* Returned hash id for keeping consistent order

* Updated dependencies

* Reworked workload test

* Perf improvement for job queue

* More relaxed behavior of internal connection pool

* Test project made dependent on env variable

* Removed id field from hash table

* Added error details for DB initializer

* Removed sequence modification as it's not supported in pgsql 9.6

* Added regression test

* Cleanup

* Lowercased sql

+++ src/Hangfire.PostgreSql/Entities/SqlJob.cs
@@ -6,14 +6,12 @@ namespace Hangfire.PostgreSql.Entities
     [UsedImplicitly(ImplicitUseTargetFlags.WithMembers)]
     internal class SqlJob
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string InvocationData { get; set; }
         public string Arguments { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime? ExpireAt { get; set; }
-
         public DateTime? FetchedAt { get; set; }
-
         public string StateName { get; set; }
         public string StateReason { get; set; }
         public string StateData { get; set; }
